Create W3C standards mixin
